New capabilities and scripts fixes.

+++ src/pyutilities/io/io_utils.py
@@ -145,30 +145,34 @@ def file_size_str(file_path: str) -> str:
     return "unknown"
 
 
-def clear_directory_except_files(directory_path: str, files_to_keep: list[str], trace: bool = False):
+def clear_directory_except_files(directory_path: str, files_to_keep: list[str], trace: bool = False) -> int:
     """Cleanup folder <directory_path>, keeping files by list <files_to_keep>. Subdirectories are
     not affected - they won't be removed (files in the subdirectories won't be touched as well).
     Usage:
         clear_directory_except_files("example_folder", ["keep_files.txt", "another_file.txt"])
     """
 
-    if not os.path.exists(directory_path):
-        log.debug(f"{directory_path=} doesn't exist!")
-        return
+    if not os.path.exists(directory_path) or not os.path.isdir(directory_path):
+        log.debug("Path %s doesn't exist or not a directory!", directory_path)
+        return 0
 
+    counter: int = 0
     for filename in os.listdir(directory_path):
         file_path = os.path.join(directory_path, filename)
         if os.path.isfile(file_path) and filename not in files_to_keep:
             try:
                 os.remove(file_path)
+                counter += 1
                 if trace:
-                    log.debug(f"{file_path=} deleted OK.")
+                    log.debug("File %s deleted OK.", file_path)
             except FileNotFoundError:
-                log.debug(f"{file_path=} was already removed.")
+                log.debug("File %s was already removed.", file_path)
             except PermissionError:
-                log.error(f"Permission denied when trying to delete {file_path=}.")
+                log.error("Permission denied when trying to delete %s!", file_path)
             except OSError as e:
-                log.exception(f"Error deleting {file_path=}: {e}")
+                log.exception("Error deleting file %s. Error: %s.", file_path, e)
+
+    return counter
 
 
 def clear_directory_except_files_recursive(directory_path, files_to_keep):
